requirements.txt
tensorflow
streamlit
pillow
opencv-python
numpy
scikit-learn